shopinvader_api_cart: extract common from existing test
